Moved algebra-ast functions to own file
